requirements.txt
langchain-groq
langchain
python-dotenv
langchain_community
langchain_core
streamlit
